+++ src/main/java/ru/javaprojects/projector/reference/technologies/TechnologyService.java
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 import ru.javaprojects.projector.common.error.IllegalRequestDataException;
 import ru.javaprojects.projector.common.error.NotFoundException;
-import ru.javaprojects.projector.common.to.FileTo;
 import ru.javaprojects.projector.common.util.FileUtil;
 import ru.javaprojects.projector.reference.technologies.model.Technology;
 
@@ -18,7 +17,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import static ru.javaprojects.projector.common.util.FileUtil.normalizePath;
 import static ru.javaprojects.projector.reference.technologies.TechnologyUtil.createNewFromTo;
 import static ru.javaprojects.projector.reference.technologies.TechnologyUtil.updateFromTo;
 
@@ -28,7 +26,7 @@ public class TechnologyService {
     private final TechnologyRepository repository;
 
     @Value("${content-path.technologies}")
-    private String contentPath;
+    private String technologyFilesPath;
 
     public Technology get(long id) {
         return repository.getExisted(id);
@@ -39,14 +37,18 @@ public class TechnologyService {
     }
 
     public Page<Technology> getAll(Pageable pageable) {
+        Assert.notNull(pageable, "pageable must not be null");
         return repository.findAllByOrderByName(pageable);
     }
 
     public Page<Technology> getAll(Pageable pageable, String keyword) {
+        Assert.notNull(pageable, "pageable must not be null");
+        Assert.notNull(keyword, "keyword must not be null");
         return repository.findAllByKeyword(keyword, pageable);
     }
 
     public Technology getByName(String name) {
+        Assert.notNull(name, "name must not be null");
         return repository.findByNameIgnoreCase(name)
                 .orElseThrow(() -> new NotFoundException("Not found technology with name =" + name, "notfound.technology",
                         new Object[]{name}));
@@ -59,8 +61,9 @@ public class TechnologyService {
             throw new IllegalRequestDataException("Technology logo file is not present",
                     "technology.logo-not-present", null);
         }
-        Technology technology = repository.saveAndFlush(createNewFromTo(technologyTo, contentPath));
-        uploadImage(technologyTo, technology.getName());
+        Technology technology = repository.saveAndFlush(createNewFromTo(technologyTo, technologyFilesPath));
+        FileUtil.upload(technologyTo.getLogo(), technologyFilesPath + FileUtil.normalizePath(technology.getName() + "/"),
+                FileUtil.normalizePath(technologyTo.getLogo().getRealFileName()));
         return technology;
     }
 
@@ -70,22 +73,16 @@ public class TechnologyService {
         Technology technology = get(technologyTo.getId());
         String oldName = technology.getName();
         String oldLogoFileLink = technology.getLogo().getFileLink();
-        repository.saveAndFlush(updateFromTo(technology, technologyTo, contentPath));
-        if (!technologyTo.getLogo().isEmpty()) {
-            uploadImage(technologyTo, technology.getName());
+        repository.saveAndFlush(updateFromTo(technology, technologyTo, technologyFilesPath));
+        if (technologyTo.getLogo() != null && !technologyTo.getLogo().isEmpty()) {
+            FileUtil.upload(technologyTo.getLogo(), technologyFilesPath + FileUtil.normalizePath(technology.getName() + "/"),
+                    FileUtil.normalizePath(technologyTo.getLogo().getRealFileName()));
             if (!oldLogoFileLink.equalsIgnoreCase(technology.getLogo().getFileLink())) {
                 FileUtil.deleteFile(oldLogoFileLink);
             }
         } else if (!technology.getName().equalsIgnoreCase(oldName)) {
-            FileUtil.moveFile(oldLogoFileLink, contentPath + FileUtil.normalizePath(technology.getName()));
+            FileUtil.moveFile(oldLogoFileLink, technologyFilesPath + FileUtil.normalizePath(technology.getName()));
         }
-    }
-
-    private void uploadImage(TechnologyTo technologyTo, String technologyName) {
-        FileTo logo = technologyTo.getLogo();
-        String fileName = normalizePath(logo.getInputtedFile() != null ? logo.getInputtedFile().getOriginalFilename() :
-                logo.getFileName());
-        FileUtil.upload(logo, contentPath + FileUtil.normalizePath(technologyName) + "/", fileName);
     }
 
     @Transactional
@@ -97,6 +94,7 @@ public class TechnologyService {
     }
 
     public Set<Technology> getAllByIds(Set<Long> ids) {
+        Assert.notNull(ids, "ids must not be null");
         return new HashSet<>(repository.findAllById(ids));
     }
 }
